Fix failing certificate renewal.
Proxying to the cerbot standalone web server was misconfigured in HAProxy.
The proxy destination was 'localhost', which is invalid inside the container.
Switched to ip-172-31-46-101.us-east-2.compute.internal like the CouchDB proxy uses.

The initial setup of the http01 challenge port was also broken.
Since HAProxy occupies port 80, we configured cerbot to start its server on a non-standard port and
proxied ACME challenge requests to it from HAProxy. Since HAProxy will fail to start if its certificate
file does not exist, certbot has to initially be launched in a standard configuration (standalone
webserver/http01_port on port 80) to acquire the first cert.

The certbot setup script assumed that 'http01_port =' would be present in the configuration after
this initial launch, but certbot omits it if the default (80) is used. Thus, the http01_port
replacement command did nothing and certbot was left to fight for port 80 with HAProxy.

http01_port is now added, if not already present, after the initial certbot run.

+++ scripts/certbot-renew-hook.sh
@@ -21,26 +21,30 @@ chmod 600 "combined.pem"
 echo "Updating certs in HAProxy..."
 echo
 
-# https://www.haproxy.com/documentation/hapee/2-3r1/management/starting-stopping/#reload-the-configuration
-docker exec proxy bash -c 'kill -SIGUSR2 $(cat /run/haproxy.pid)' || echo "Could not update live certs"
+if docker ps | grep "proxy"; then
+    # https://www.haproxy.com/documentation/hapee/2-3r1/management/starting-stopping/#reload-the-configuration
+    docker exec proxy bash -c 'kill -SIGUSR2 $(cat /run/haproxy.pid)' || echo "Could not update live certs"
 
-# Can also use the runtime API to update the cert but it's a bit more complicated/fragile for our setup
-# Have to add:
-#
-# global
-#  # Enable the runtime API so we can reload SSL certs
-#  stats socket /run/haproxy-api.sock mode 600 level admin
-#
-# to haproxy.cfg
-# https://www.haproxy.com/blog/dynamic-ssl-certificate-storage-in-haproxy/
-#
-# API="nc -U /run/haproxy-api.sock"
-# CERT_CONTENTS=$(cat combined.pem)
-# # docker exec proxy bash -c "echo 'show ssl cert' | ${API}"
-# # docker exec proxy bash -c "echo 'show ssl cert /certs/combined.pem' | ${API}"
-# docker exec proxy bash -c "echo -e 'set ssl cert /certs/combined.pem <<\n${CERT_CONTENTS}\n' | ${API}"
-# # docker exec proxy bash -c "echo 'show ssl cert' | ${API}"
-# # docker exec proxy bash -c "echo 'show ssl cert */etc/haproxy/certs/site.pem' | ${API}"
-# docker exec proxy bash -c "echo 'commit ssl cert /etc/haproxy/certs/site.pem' | ${API}"
+    # Can also use the runtime API to update the cert but it's a bit more complicated/fragile for our setup
+    # Have to add:
+    #
+    # global
+    #  # Enable the runtime API so we can reload SSL certs
+    #  stats socket /run/haproxy-api.sock mode 600 level admin
+    #
+    # to haproxy.cfg
+    # https://www.haproxy.com/blog/dynamic-ssl-certificate-storage-in-haproxy/
+    #
+    # API="nc -U /run/haproxy-api.sock"
+    # CERT_CONTENTS=$(cat combined.pem)
+    # # docker exec proxy bash -c "echo 'show ssl cert' | ${API}"
+    # # docker exec proxy bash -c "echo 'show ssl cert /certs/combined.pem' | ${API}"
+    # docker exec proxy bash -c "echo -e 'set ssl cert /certs/combined.pem <<\n${CERT_CONTENTS}\n' | ${API}"
+    # # docker exec proxy bash -c "echo 'show ssl cert' | ${API}"
+    # # docker exec proxy bash -c "echo 'show ssl cert */etc/haproxy/certs/site.pem' | ${API}"
+    # docker exec proxy bash -c "echo 'commit ssl cert /etc/haproxy/certs/site.pem' | ${API}"
+else
+    echo "Proxy container was not active."
+fi
 
 cd -
